add unit test for vue and agular
